+++ src/config/aboutConfig.js
@@ -1,0 +1,93 @@
+const aboutConfig = {
+    aboutText1 : `I'm a DevOps Engineer from Delhi, India, working in the Software Engineering field and web development.
+          Streamlining infrastructure automation and ensuring seamless CI/CD pipelines, I empower development teams
+          with reliable, scalable, and efficient deployment solutions.`,
+    
+    aboutText2 : `My job is to build your website so that it is functional and user-friendly but at the same time attractive.
+          Moreover, I add personal touch to your product and make sure that is eye-catching and easy to use. My aim is
+          to bring across your message and identity in the most creative way. I created web design for many famous
+          brand companies.`,
+    
+    serviceList : [
+        {
+            title: 'DevOps',
+            description: 'Efficient and Stable Pipeline and Script made at a professional level.',
+            icon: 'icon-1',
+        },
+        {
+            title: 'Cloud',
+            description: 'Robust and scalable cloud architecture crafted with expertise, leveraging AWS services for optimal performance and reliability.',
+            icon: 'icon-2',
+        },
+        {
+            title: 'Data Structure and Algorithms',
+            description: 'Mastering core programming languages and algorithms to deliver high-performance, efficient, and innovative software solutions.',
+            icon: 'icon-3',
+        },
+        {
+            title: 'Development',
+            description: 'High-quality development of APIs at the professional level.',
+            icon: 'icon-4',
+        },
+    ],
+    testimonialsList: [
+        {
+            avatar: 'avatar-1',
+            name: 'Daniel Lewis',
+            testimony: `Richard was hired to create a corporate identity. We were very pleased with the work done. She has a lot of
+            experience and is very concerned about the needs of client.`,
+        },
+        {
+            avatar: 'avatar-2',
+            name: 'Jessica Miller',
+            testimony: `Richard was hired to create a corporate identity. We were very pleased with the work done. She has a lot of
+            experience and is very concerned about the needs of client.`,
+        },
+        {
+            avatar: 'avatar-3',
+            name: 'Emily Evans',
+            testimony: `Richard was hired to create a corporate identity. We were very pleased with the work done. She has a lot of
+            experience and is very concerned about the needs of client.`,
+        },
+        {
+            avatar: 'avatar-4',
+            name: 'Henry William',
+            testimony: `Richard was hired to create a corporate identity. We were very pleased with the work done. She has a lot of
+            experience and is very concerned about the needs of client.`,
+        },
+    ],
+    clientsList:[
+        {
+            name:'default',
+            logo:'logo-1-color.png',
+            link: 'https://www.google.com',
+        },
+        {
+            name:'default',
+            logo:'logo-2-color.png',
+            link: 'https://www.google.com',
+        },
+        {
+            name:'default',
+            logo:'logo-1-color.png',
+            link: 'https://www.google.com',
+        },
+        {
+            name:'default',
+            logo:'logo-1-color.png',
+            link: 'https://www.google.com',
+        },
+        {
+            name:'default',
+            logo:'logo-1-color.png',
+            link: 'https://www.google.com',
+        },
+        {
+            name:'default',
+            logo:'logo-1-color.png',
+            link: 'https://www.google.com',
+        },
+    ],
+}
+
+export default aboutConfig;