+++ src/config/careerConfig.js
@@ -1,0 +1,44 @@
+export const educationConfig = [
+  {
+    org: "Maharaja Surajmal Institute Of Technology",
+    session: "2019 — 2022",
+    course: "Bachelor of Technology in Information Technology",
+  },
+  {
+    org: "Pusa Institute Of Technology",
+    session: "2014 — 2017",
+    course: "Diploma in Mechanical Engineering",
+  },
+  {
+    org: "Deepanshu Public School",
+    session: "2013 — 2014",
+    course: "Secondary School Certificate (Class 10th)",
+  },
+];
+
+export const experienceConfig = [
+  {
+    designation: "Associate Engineer, DevOps and Cloud Technologies",
+    tenure: "Oct, 2022 — Present",
+    name: "Impressico Business Solutions",
+  },
+];
+
+export const skillsConfig = [
+  {
+    name: "DevOps",
+    level: "80",
+  },
+  {
+    name: "Cloud",
+    level: "70",
+  },
+  {
+    name: "Coding Languages (Python, Java, Groovy, JavaScript)",
+    level: "90",
+  },
+  {
+    name: "Development",
+    level: "60",
+  },
+];