+++ src/config/navConfig.js
@@ -1,0 +1,28 @@
+const navComponents = [
+    {
+      page: "About",
+      to: "/",
+    },
+    {
+      page: "Carrer",
+      to: "/career",
+    },
+    {
+      page: "Resume",
+      to: "/resume",
+    },
+    {
+      page: "Portfolio",
+      to: "/portfolio",
+    },
+    {
+      page: "Blog",
+      to: "/blog",
+    },
+    {
+      page: "Contact",
+      to: "/contact",
+    },
+  ];
+
+export default navComponents;