+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 ## Changelog
 
+**Branch**: `release/2.0.0`
+
+### Version 2.0.0 - Initial Release
+
+#### Features:
+- **Added Portfolio Website v2.0.0 Code**:
+  - added code for second portfolio Website
+
+---
+
 **Branch**: `master`
 
 ### Version 1.0.0 - Initial Release